Completed all data sorting functions.
In a big milestone, our functions can now take
a dataframe full of unsorted information and
return organized, graphable data. The final
dataset can be understood as four 3D arrays (one
for each region) in which the height represents
all disaster types, the width represents their
associated costs and deaths in a year, and the depth
represents all of the years in which data was taken
for that particular US region.